requirements-dev.txt
-r requirements.txt
codecov
fast-alchemy
flake8

Flask-Restless==0.17.0
flask-restless-datamodel==0.1.7
isort


pre-commit==1.15.2

pytest >= 4.5.0
pytest-cov
pytest-xdist
requests-flask-adapter
yapf
